+++ sites/geohub/src/routes/api/stac/catalog/[id]/item/+server.ts
@@ -1,15 +1,20 @@
 import { AccessLevel } from '$lib/config/AppConfig';
 import type { RequestHandler } from './$types';
 import { env } from '$env/dynamic/private';
-import { createDatasetLinks, getSTACs } from '$lib/server/helpers';
+import {
+	createDatasetLinks,
+	createMosaicDataSetFeature,
+	createTitilerMosaicJsonEndpoint,
+	getSTACs
+} from '$lib/server/helpers';
 import { error } from '@sveltejs/kit';
 import type { DatasetFeature, Stac, StacCollection, StacItemFeature, Tag } from '$lib/types';
 import { generateHashKey, resolveRelativeUrl } from '$lib/helper';
 
 export const GET: RequestHandler = async ({ params, url }) => {
-	const type = params.type;
+	const id = params.id;
 	const stacCatalogs = await getSTACs('catalog');
-	const stac = stacCatalogs.find((x) => x.id === type);
+	const stac = stacCatalogs.find((x) => x.id === id);
 	if (!stac) {
 		throw error(
 			400,
@@ -17,11 +22,11 @@ export const GET: RequestHandler = async ({ params, url }) => {
 		);
 	}
 
-	const itemUrl = url.searchParams.get('url');
+	const itemUrls = url.searchParams.getAll('url');
 	const assetName = url.searchParams.get('asset');
 	const collectionUrl = url.searchParams.get('collection');
 
-	if (!itemUrl) {
+	if (itemUrls.length === 0) {
 		throw error(400, { message: `url query param for STAC item URL is missing.` });
 	}
 	if (!assetName) {
@@ -33,16 +38,44 @@ export const GET: RequestHandler = async ({ params, url }) => {
 		});
 	}
 
-	const res = await fetch(itemUrl);
-	const itemFeature: StacItemFeature = await res.json();
+	let feature: DatasetFeature;
 
-	const feature = await generateDataSetFeature(
-		stac,
-		itemFeature,
-		assetName,
-		itemUrl,
-		collectionUrl
-	);
+	const res = await fetch(collectionUrl);
+	const collection: StacCollection = await res.json();
+
+	if (itemUrls.length === 1) {
+		// single item
+		const itemUrl = itemUrls[0];
+		const res = await fetch(itemUrl);
+		const itemFeature: StacItemFeature = await res.json();
+
+		feature = await getDataSetFeature(stac, itemFeature, assetName, itemUrl, collection);
+	} else {
+		// mosaicjson
+		const features: DatasetFeature[] = [];
+
+		const sortedItems = itemUrls.sort();
+
+		for (const itemUrl of sortedItems) {
+			const res = await fetch(itemUrl);
+			const itemFeature: StacItemFeature = await res.json();
+
+			feature = await getDataSetFeature(stac, itemFeature, assetName, itemUrl, collection);
+			features.push(feature);
+		}
+		const urls: string[] = features.map((f) => f.properties.url);
+		const ids: string[] = features.map((f) => f.properties.id.replace('/', '_')).sort();
+		const name = `${id}/${collection.id}/${ids.join('/')}/${assetName}/mosaicjson.json`;
+		const mosaicjson = await createTitilerMosaicJsonEndpoint(urls, name);
+		const mosaicjsonFeature = await createMosaicDataSetFeature(features, mosaicjson);
+		mosaicjsonFeature.properties = createDatasetLinks(
+			mosaicjsonFeature,
+			url.origin,
+			env.TITILER_ENDPOINT
+		);
+		return new Response(JSON.stringify(mosaicjsonFeature));
+	}
+
 	feature.properties = createDatasetLinks(feature, url.origin, env.TITILER_ENDPOINT);
 
 	const selfLink = feature.properties.links.find((l) => l.rel === 'self');
@@ -51,18 +84,15 @@ export const GET: RequestHandler = async ({ params, url }) => {
 	return new Response(JSON.stringify(feature));
 };
 
-const generateDataSetFeature = async (
+const getDataSetFeature = async (
 	stac: Stac,
 	item: StacItemFeature,
 	assetName: string,
 	itemUrl: string,
-	collectionUrl: string
+	collection: StacCollection
 ) => {
 	const assetItem = item.assets[assetName];
 	const assetUrl = resolveRelativeUrl(assetItem.href, itemUrl);
-
-	const res = await fetch(collectionUrl);
-	const collection: StacCollection = await res.json();
 
 	let providers: Tag[] = collection.providers?.map((p) => {
 		return { key: 'provider', value: p.name };